Prime generation logic
The meat of the problem is here. I have based this first version on the sieve of Eratostenes, but due that we request a number of primes, and we don't know how many test are required to get that number, the order of operations has to change on regards to the standard algorithm. Stream is used to provide this lazy, infinite loop. Due to how most functional languages work (lack of state, parameters are the way to control results), the different subparts have been tried on iex, before being added into the file all together.