More renaming and refactoring
In the previous renaming I didn't correct mix.ex, which has been done now. Then, through the marvels of Wikipedia I discovered that what I was doing was really a TrialDivision sieve, not an infinite Sieve of Erathostenes. So names have been changed, and I have move the code into a Sieves subfolder, so when adding a new one makes it better structured